+++ test/test.cc
@@ -3,6 +3,7 @@
 #include <gtest/gtest.h>
 
 #include <onmt/Tokenizer.h>
+#include <onmt/Alphabet.h>
 
 using namespace onmt;
 
@@ -141,6 +142,16 @@ TEST(TokenizerTest, SegmentNumbers) {
   test_tok_and_detok(tokenizer,
                      "1984 mille neuf cent quatrevingt-quatre",
                      "1￭ 9￭ 8￭ 4 mille neuf cent quatrevingt ￭-￭ quatre");
+}
+
+TEST(TokenizerTest, GetAlphabet) {
+  for (const auto& a : onmt::alphabet_ranges) {
+    const auto& range = a.first;
+    const auto& name = a.second;
+
+    EXPECT_EQ(name, onmt::get_alphabet(range.first));
+    EXPECT_EQ(name, onmt::get_alphabet(range.second));
+  }
 }
 
 TEST(TokenizerTest, SegmentAlphabet) {