+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes and improvements
 
 * Fix incorrect behavior when `--segment_alphabet` is not set by the client
+* Fix alphabet identification
 
 ## [v1.0.0](https://github.com/OpenNMT/Tokenizer/releases/tag/v1.0.0) (2017-12-11)